ext-source/docs-notebooks/voyageai/apps/rag-in-memory/requirements.txt
voyageai
numpy
langchain-community
langchain-text-splitters
pypdf
python-dotenv
anthropic
openai